@@ -1,10 +1,9 @@
 module example.persistence {
-    requires java.persistence;
+    requires jakarta.persistence;
     requires java.sql;
 
-    requires java.xml.bind;
-    requires net.bytebuddy;
-
+    requires org.hibernate.orm.core;
+    requires spring.core;
     requires spring.context;
     requires spring.data.jpa;
     requires spring.jdbc;
@@ -12,7 +11,7 @@ module example.persistence {
     requires spring.tx;
 
     exports red.jackal.training.spring.jpms.entity;
-    exports red.jackal.training.spring.jpms.repository to example.service;
+    exports red.jackal.training.spring.jpms.repository to example.service, example.web;
 
     opens red.jackal.training.spring.jpms.config to spring.core, spring.beans, spring.context;
     opens red.jackal.training.spring.jpms.entity to org.hibernate.orm.core, spring.core;
