@@ -1,5 +1,6 @@
 package red.jackal.training.spring.jpms.config;
 
+import jakarta.persistence.EntityManagerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
@@ -11,11 +12,10 @@ import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
-import javax.persistence.EntityManagerFactory;
 import javax.sql.DataSource;
 
 @Configuration
-@EnableJpaRepositories(basePackages = { "red.jackal.training.spring.jpms.repository" })
+@EnableJpaRepositories(basePackages = {"red.jackal.training.spring.jpms.repository"})
 @EnableTransactionManagement
 public class DataConfig {
 
